move tests, uncomment consistency test
